Move out free functions (arr1 etc)
